♻️  Header 컴포넌트 분리
 - issue #6
