+++ src/components/Text.jsx
@@ -1,4 +1,3 @@
-/* eslint-disable no-console */
 import React from 'react';
 import styled from 'styled-components';
 import THEME from 'variable/theme';
